requirements.txt
flask==1.1.4
flask-restx==0.4.0
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.8.6
pytest==6.2.4
pytest-cov==2.12.0
flake8==3.9.2